Fixed attachement file size type

+++ pkg/infra/models/jira_issue_attachments.go
@@ -13,7 +13,7 @@ type IssueAttachmentScheme struct {
 	Filename  string      `json:"filename,omitempty"`  // The filename of the attachment.
 	Author    *UserScheme `json:"author,omitempty"`    // The author of the attachment.
 	Created   string      `json:"created,omitempty"`   // The creation time of the attachment.
-	Size      int         `json:"size,omitempty"`      // The size of the attachment.
+	Size      int64       `json:"size,omitempty"`      // The size of the attachment.
 	MimeType  string      `json:"mimeType,omitempty"`  // The MIME type of the attachment.
 	Content   string      `json:"content,omitempty"`   // The content of the attachment.
 	Thumbnail string      `json:"thumbnail,omitempty"` // The thumbnail of the attachment.
@@ -26,7 +26,7 @@ type IssueAttachmentMetadataScheme struct {
 	Filename  string      `json:"filename,omitempty"`  // The filename of the attachment.
 	Author    *UserScheme `json:"author,omitempty"`    // The author of the attachment.
 	Created   string      `json:"created,omitempty"`   // The creation time of the attachment.
-	Size      int         `json:"size,omitempty"`      // The size of the attachment.
+	Size      int64       `json:"size,omitempty"`      // The size of the attachment.
 	MimeType  string      `json:"mimeType,omitempty"`  // The MIME type of the attachment.
 	Content   string      `json:"content,omitempty"`   // The content of the attachment.
 	Thumbnail string      `json:"thumbnail,omitempty"` // The thumbnail of the attachment.
